+++ site/tmpl/recipes/default.php
@@ -17,6 +17,8 @@ use \Joomla\CMS\Language\Text;
 use \Joomla\CMS\Layout\LayoutHelper;
 use \Joomla\CMS\Session\Session;
 use \Joomla\CMS\User\UserFactoryInterface;
+use Webtest\Component\Web357test\Site\Enum\RecipeDifficulty;
+use Webtest\Component\Web357test\Site\Enum\ServingSize;
 
 HTMLHelper::_('bootstrap.tooltip');
 HTMLHelper::_('behavior.multiselect');
@@ -44,12 +46,12 @@ $wa->useStyle('com_web357test.list');
 <?php endif;?>
 <form action="<?php echo htmlspecialchars(Uri::getInstance()->toString()); ?>" method="post"
 	  name="adminForm" id="adminForm">
-	
+
 	<div class="table-responsive">
 		<table class="table table-striped" id="recipeList">
 			<thead>
 			<tr>
-				
+
 					<th class=''>
 						<?php echo HTMLHelper::_('grid.sort',  'COM_WEB357TEST_RECIPES_ID', 'a.id', $listDirn, $listOrder); ?>
 					</th>
@@ -78,6 +80,10 @@ $wa->useStyle('com_web357test.list');
 						<?php echo HTMLHelper::_('grid.sort',  'COM_WEB357TEST_RECIPES_DIFFICULTY', 'a.difficulty', $listDirn, $listOrder); ?>
 					</th>
 
+					<th class=''>
+						<?php echo HTMLHelper::_('grid.sort',  'COM_WEB357TEST_FORM_LBL_SERVING_SIZE', 'a.serving_size', $listDirn, $listOrder); ?>
+					</th>
+
 						<?php if ($canEdit || $canDelete): ?>
 					<th class="center">
 						<?php echo Text::_('COM_WEB357TEST_RECIPES_ACTIONS'); ?>
@@ -103,7 +109,7 @@ $wa->useStyle('com_web357test.list');
 				<?php endif; ?>
 
 				<tr class="row<?php echo $i % 2; ?>">
-					
+
 					<td>
 						<?php echo $item->id; ?>
 					</td>
@@ -136,8 +142,23 @@ $wa->useStyle('com_web357test.list');
 						<?php echo $item->cooking_time; ?>
 					</td>
 					<td>
-						<?php echo $item->difficulty; ?>
+                        <?php
+                        $difficulty = RecipeDifficulty::tryFrom($item->difficulty);
+                        if ($difficulty)
+                        {
+                            echo $difficulty->getHtml();
+                        }
+                        ?>
 					</td>
+                    <td>
+                        <?php
+                        $servingSize = ServingSize::tryFrom($item->serving_size);
+                        if ($servingSize)
+                        {
+                            echo $servingSize->getText();
+                        }
+                        ?>
+                    </td>
 					<?php if ($canEdit || $canDelete): ?>
 						<td class="center">
 							<?php $canCheckin = Factory::getApplication()->getIdentity()->authorise('core.manage', 'com_web357test.' . $item->id) || $item->checked_out == Factory::getApplication()->getIdentity()->id; ?>