+++ CHANGELOG.md
@@ -8,3 +8,8 @@ All notable changes to this project will be documented in this file.
 
 - New parameter "Serving size" for Recipe in the administrator form
 - New parameter "Serving size" for Recipe in frontend form
+
+### Changed
+
+- Use Font Awesome icons for difficulty levels in the frontend recipe list and single page view instead of plain text
+- Display serving size in the frontend recipe list and single page view